Changed float64 to float32 for SAC compatibility

+++ melgym/envs/hvac.py
@@ -37,14 +37,14 @@ class EnvHVAC(Env):
         low_obs = np.zeros(n_obs)
         high_obs = np.inf * np.ones(n_obs)
         self.observation_space = spaces.Box(
-            low=low_obs, high=high_obs, dtype=np.float64)
+            low=low_obs, high=high_obs, dtype=np.float32)
 
         # Action space
         # low_act = np.zeros(n_actions)
         low_act = max_vel * -np.ones(n_actions)
         high_act = max_vel * np.ones(n_actions)
         self.action_space = spaces.Box(
-            low=low_act, high=high_act, dtype=np.float64)
+            low=low_act, high=high_act, dtype=np.float32)
 
         # self.action_space = spaces.Discrete(3)
 
@@ -121,7 +121,7 @@ class EnvHVAC(Env):
         self.last_velocity = 0
         self.last_pressures = {}
 
-        return np.array(obs, dtype=np.float64), info
+        return np.array(obs, dtype=np.float32), info
 
     def step(self, action):
         """
@@ -183,7 +183,7 @@ class EnvHVAC(Env):
         self.last_distances = distances
         self.last_truncated = truncated
 
-        return np.array(obs, np.float64), reward, terminated, truncated, info
+        return np.array(obs, np.float32), reward, terminated, truncated, info
 
     def render(self, time_bt_frames=0.1):
         """
@@ -280,7 +280,7 @@ class EnvHVAC(Env):
 
             record['time'] = float(last_line[0])
             for i, cv_id in enumerate(self.__get_edf_cvs()):
-                record[cv_id] = np.float64(last_line[i+1])
+                record[cv_id] = np.float32(last_line[i+1])
 
         return record
